use smaller index types for fuzzing
this *does* restrict the size of input that we're able to test with, but
the smaller numbers are a *lot* more manageable.

+++ fuzz/fuzz_targets/no_features_basic_slices.rs
@@ -24,11 +24,10 @@ impl<'d> Arbitrary<'d> for UpperLetter {
 }
 
 fuzz_target!(
-    |cmds: CommandSequence<BasicCommand<usize, Constant<UpperLetter>>>| {
+    |cmds: CommandSequence<BasicCommand<u8, Constant<UpperLetter>>>| {
         let cmds = cmds.map(|cmd| cmd.map_slice(|c| Constant(c.0)));
 
-        let mut runner: RunnerState<usize, Constant<UpperLetter>, NoFeatures, 3> =
-            RunnerState::init();
+        let mut runner: RunnerState<u8, Constant<UpperLetter>, NoFeatures, 3> = RunnerState::init();
 
         for c in cmds.cmds {
             runner.run_basic_cmd(&c);
